дз модуль 5 задание 1
