+++ index.html
@@ -234,7 +234,7 @@
       </section>
     </template>
 
-    <script src="./js/main.js" defer></script>
+    <script src="./js/main.js" type="module" defer></script>
     <script src="./js/functions.js" defer></script>
 </body>
 </html>
